Add comment about risk of component unmounting
This is in response to the question:

Imagine this is an app with multiple pages, upon navigating to this page this effect is triggered. Do you see any issue with navigating to this page and then quickly navigating to another page before the api responds? How would you solve it?